# PERF: DataFrame.first_valid_index/last_valid_index for EA dtypes