@jbrockmendel 2023-02-22
wow thats a huge difference.  is it all in the len(values) or something?