@lukemanley 2023-02-22
> wow thats a huge difference. is it all in the len(values) or something?

`self._values` on a multi-column EA-backed frame converts to an object dtype ndarray. This avoids the going through the object dtype conversion.