@lukemanley 2023-02-22
> This is basically the same as with MultiIndex._values?

similar in the sense they're both expensive conversions through object dtype. `MultiIndex._values` has the additional overhead of wrapping in tuples.