@phofl 2023-02-22
Yep, thx, that's what I was thinking